backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit==1.28.1
matplotlib==3.8.2
numpy==1.24.3
